# lib/api-zod/src/generated/types/analyzeCompetitorsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeCompetitorsBody {
  /**
   * List of 3-5 FMCG company names to analyze
   * @minItems 3
   * @maxItems 5
   */
  companies: string[];
}