# lib/api-zod/src/generated/types/intelligenceReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CompanyProfile } from "./companyProfile";

export interface IntelligenceReport {
  id: number;
  companies: string[];
  profiles: CompanyProfile[];
  createdAt: Date;
  title: string;
}