# lib/api-zod/src/generated/types/sourcedDataPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SourcedDataPoint {
  /** The data point value */
  value: string;
  /** Source URL for this data point */
  source: string;
  /** Title or description of the source */
  sourceTitle?: string;
}